VB overlap reseg more fully functional

+++ egs/jsalt2019-diadet/v4/run_028_diar_overlap_assignment.sh
@@ -1,9 +1,13 @@
 #!/bin/bash
 # Copyright
-#                2019   Latané Bullock (JSALT 2019)
+#            
 #
 # Apache 2.0.
 #
+#
+#
+# JSALT 2019, Latané Bullock
+
 
 
 . ./cmd.sh
@@ -18,24 +22,28 @@ config_file=default_config.sh
 
 
 ovl_out_dir=exp/overlap_models/test_raw
-vad_out_dir=exp/vad
-vb_out_dir=
+vb_suff=VB
 ovl_suff=OVLassign
 
 score_dir=exp/diarization/$nnet_name/$be_diar_name
 
-# retrieves all dsets which have had VB resegmentation performed
+# Retrieves all dsets which we want to perform resegmentation on
+# We first get all dsets which would have a corresponding VB reseg
+# Can perform filtering with inverse grep if needed
 dsets_path=$score_dir
-dsets=`find $dsets_path  -maxdepth 1 -name "jsalt19_spkdiar*VB" | xargs -l basename | sort`
+dsets=`find $dsets_path  -maxdepth 1 -name "jsalt19_spkdiar*" \
+  | xargs -l basename \
+  | sort \
+  | grep -v VB \
+  | grep -v sri \
+  `
 
-# TESTING(FIXME)
-dsets="jsalt19_spkdiar_ami_dev_Mix-Headset_VB"
+
 
 if [ $stage -le 1 ]; then
-	# Given an overlap RTTM, VAD rttm, and the VB-HMM-GMM q matrix (speaker poster attributions), 
-	# we perform frame-level reassignment to two speakers in overlap segments
 
 	for dset in $dsets; do
+		
 		# We need to do some name parsing
 		if [[ "$dset" =~ .*ami.* ]];then
       		corp=ami
@@ -57,48 +65,56 @@ if [ $stage -le 1 ]; then
 
 		# TESTING(FIXME) need to add the rest of the mics
 		if [[ "$dset" =~ .*Array1.* ]];then
-      		mic=Array1-01
+      		mic='_Array1-01'
     	elif [[ "$dset" =~ .*Array2.* ]];then
-      		mic=Array2-01
+      		mic='_Array2-01'
     	elif [[ "$dset" =~ .*Mix-Headset.* ]];then
-      		mic=Mix-Headset
-		else echo "dset -- $dset -- unable to be parsed"; exit 1; 
+      		mic='_Mix-Headset'
+    	elif [[ "$dset" =~ .*U01.* ]];then
+      		mic='_U01'
+    	elif [[ "$dset" =~ .*U06.* ]];then
+      		mic='_U06'
+		else mic='';
+		fi
+
+		dset_vb=${dset}_${vb_suff}
+		if [ ! -d ${dsets_path}/${dset_vb} ]; then 
+			continue
 		fi
 		
-		
 		# This is the dir that the overlap assignment will use 
-		output_dir=${dsets_path}/${dset}_${ovl_suff}
+		output_dir=${dsets_path}/${dset_vb}_${ovl_suff}
 		
 		mkdir -p $output_dir
 
-		#TESTING(FXIME)
+		#TESTING(FXIME)-clear dir before starting
 		rm -rf $output_dir/tmp
 		rm -rf $output_dir/*.log
 		rm -rf $output_dir/result.*
 		rm -rf $output_dir/rttm
 
 		# get rttm from BEFORE VB reseg
-		# cp $dsets_path/$dset/rttm $output_dir/rttm_in
+		cp $dsets_path/$dset/plda_scores_tbest/rttm $output_dir/rttm_in
 		
-		overlap_rttm=${ovl_out_dir}/diar_overlap_${corp}_${part}_${mic}.rttm
+		overlap_rttm=${ovl_out_dir}/diar_overlap_${corp}_${part}${mic}.rttm
 		cat $overlap_rttm > $output_dir/overlap.rttm
 		
-		# The data dir should contain a pyannoteVAD, utt2spk, and utt2num_frames which are of interest
-		data_dir=data/jsalt19_spkdiar_${corp}_${part}_${mic}
+		# The data dir should contain a utt2spk and utt2num_frames used in VB 
+		data_dir=data/jsalt19_spkdiar_${corp}_${part}${mic}
 		cp $data_dir/utt2spk $output_dir/
 		cp $data_dir/utt2num_frames $output_dir/
 		
 		# We need the q matrix for each utterance from VB resegmentation
 		mkdir -p $output_dir/q_mats
 		mkdir -p $output_dir/tmp
-		# cp $dsets_path/$dset/tmp/*.npy $output_dir/q_mats
+		cp $dsets_path/$dset_vb/tmp/*.npy $output_dir/q_mats
 		
 
 
 		# TESTING(FIXME) - this env activation should be placed more strategiclaly 
 		source activate pyannote
 		
-		# Usage: diar_ovl_assignment.py [-h] q_mat_dir ovl_rttm vad_rttm output_dir
+		# Usage: diar_ovl_assignment.py [-h] <overlap_dir>
 		# At this point, the output dir should contain the q matrix from VB reseg
 		# an overlap rttm, and a vad rttm 
 		python3 local/diar_ovl_assignment.py $output_dir
@@ -117,7 +133,6 @@ ovl_dsets=`find $score_dir  -maxdepth 1 -name "jsalt19_spkdiar*" \
   | grep "$ovl_suff" \
   `
 
-
 if [ $stage -le 2 ]; then
   
   for name in $ovl_dsets
@@ -135,7 +150,7 @@ if [ $stage -le 2 ]; then
     echo $name
 
     # TESTING(FIXME)
-    # we can skip those files which have already been evaluated
+    # we can skip those rttms which have already been evaluated
     if [ -s $score_dir/$name/result.pyannote-der ];then
       continue
     fi
